fix(unzipper): fix unzipper stream using fs-extra since 0.12.0
The unzipper dependency 0.12.0 switched to fs-extra which breaks the piping logic after an icon package was downloaded.
The read stream triggers the close event twice, thus it is stuck inside the extracting loop after the first event trigger.

Related dependencies are updated as well

Also fixed check for previously not properly extraction of zip file